requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
flask-cors==4.0.0
requests==2.31.0
flask-socketio==5.3.6
eventlet==0.33.3
python-engineio==4.8.0
python-socketio==5.10.0
